Teensy 4.0 and 4.1 support
This commit adds support for the Teensy 4.0 and 4.1 boards.

It modifies the ports/mimxrt10xx/tusb_config.h and boards.c so that
either USB port 0 or usb port 1 can be used.  This is useful for any
IMXRT1062 based board.  Though less useful for the teensy 4.0 because
it doens't have that port pinned out.